// backend/src/utils/jwt.d.ts
export declare const generateAccessToken: (userId: string, roleName: string, storeId?: string | null) => string;
export declare const generateRefreshToken: (userId: string) => string;
export declare const verifyAccessToken: (token: string) => {
    userId: string;
    role: string;
};
export declare const verifyRefreshToken: (token: string) => {
    userId: string;
};
//# sourceMappingURL=jwt.d.ts.map